Logs: describe_log_groups() now returns the logStreamArn-property
