+++ tests/test_logs/test_logs_cloudformation.py
@@ -1,33 +1,52 @@
 import json
+from uuid import uuid4
 
 import boto3
+import pytest
 
-from moto import mock_aws
+from tests import aws_verified
 
 
-@mock_aws
+@aws_verified
+@pytest.mark.aws_verified
 def test_tagging():
     logs_client = boto3.client("logs", region_name="us-east-1")
     cf_client = boto3.client("cloudformation", region_name="us-east-1")
+
+    log_group_name = f"/moto/test/{str(uuid4())}"
 
     template = {
         "AWSTemplateFormatVersion": "2010-09-09",
         "Resources": {
             "testGroup": {
                 "Type": "AWS::Logs::LogGroup",
-                "Properties": {"Tags": [{"Key": "foo", "Value": "bar"}]},
+                "Properties": {
+                    "LogGroupName": log_group_name,
+                    "Tags": [{"Key": "foo", "Value": "bar"}],
+                },
             }
         },
     }
     template_json = json.dumps(template)
+    stack_name = f"moto-test-{str(uuid4())[0:6]}"
     cf_client.create_stack(
-        StackName="test_stack",
+        StackName=stack_name,
         TemplateBody=template_json,
     )
+    waiter = cf_client.get_waiter("stack_create_complete")
+    waiter.wait(StackName=stack_name)
 
-    arn = logs_client.describe_log_groups()["logGroups"][0]["arn"]
-    tags = logs_client.list_tags_for_resource(resourceArn=arn)["tags"]
-    assert tags == {"foo": "bar"}
+    group = logs_client.describe_log_groups(logGroupNamePrefix=log_group_name)[
+        "logGroups"
+    ][0]
 
-    cf_client.delete_stack(StackName="test_stack")
-    assert logs_client.describe_log_groups()["logGroups"] == []
+    tags = logs_client.list_tags_for_resource(resourceArn=group["logGroupArn"])["tags"]
+    assert tags["foo"] == "bar"
+
+    cf_client.delete_stack(StackName=stack_name)
+    waiter = cf_client.get_waiter("stack_delete_complete")
+    waiter.wait(StackName=stack_name)
+    assert (
+        logs_client.describe_log_groups(logGroupNamePrefix=log_group_name)["logGroups"]
+        == []
+    )
